[zig_zag_scan.py]: add comments

+++ src/zig_zag_scan.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-# Time-stamp: < zig_zag_scan.py 2016-05-07 10:00:14 >
+# Time-stamp: < zig_zag_scan.py 2016-05-07 10:24:28 >
 """
 Z字形的扫描,zig-zag scan
 """ 
@@ -17,7 +17,7 @@ order_of_dct_table = np.array([
 	[20, 22, 33, 38, 46, 51, 55, 60],
 	[21, 34, 37, 47, 50, 56, 59, 61],
 	[35, 36, 48, 49, 57, 58, 62, 63],
-],dtype=np.uint8)
+])
 
 # 测试矩阵，课本P129，附有正确答案
 test_table=np.array([
@@ -37,14 +37,15 @@ def generate_dict(direction):
 	seq_dict = {}
 	if direction == 'forward':
 		for i,j in [(i,j) for i in xrange(8) for j in xrange(8)]:
-			seq_dict[order_of_dct_table[i, j]] = (i * 8 + j)
+			seq_dict[int(order_of_dct_table[i, j])] = int(i * 8 + j)
 	elif direction == 'backward':	
 		for i,j in [(i,j) for i in xrange(8) for j in xrange(8)]:
-			seq_dict[i * 8 + j] = order_of_dct_table[i, j]
+			seq_dict[int(i * 8 + j)] = int(order_of_dct_table[i, j])
 	else:
 		return
 	return seq_dict
 
+# 把8x8的矩阵按照z字形顺序转成1x64的列表
 def get_seq_1x64(input_matrix, seq_dict):
 	list_DCT = []
 	for i in xrange(64):
@@ -52,8 +53,9 @@ def get_seq_1x64(input_matrix, seq_dict):
 		list_DCT.append(input_matrix[order / 8, order % 8])
 	return list_DCT
 
+# 把1x64的列表转成8x8矩阵
 def restore_matrix_from_1x64(input_list, seq_dict):
-	output_matrix = np.zeros(64,dtype=np.uint8).reshape(8,8)
+	output_matrix = np.zeros(64).reshape(8,8)
 	for i,j in [(i,j) for i in xrange(8) for j in xrange(8)]:
 		order = seq_dict[8 * i + j]
 		output_matrix[i, j] = input_list[order]
@@ -62,12 +64,14 @@ def restore_matrix_from_1x64(input_list, seq_dict):
 def test():
 	print "original table:"
 	print test_table
-	
+
+	# 转成1x64
 	dict_forward = generate_dict('forward')
 	list_DCT_forward = get_seq_1x64(test_table, dict_forward)
 	print "forward:"
 	print list_DCT_forward
-	
+
+	# 转成8x8
 	print "backward:"
 	dict_backward = generate_dict('backward')
 	np_DCT_backward = restore_matrix_from_1x64(list_DCT_forward, dict_backward)
